Lawyers' Ratings of State Judges in the US Superior Court

The dataset name has been changed to 'USJudgeRatings_df' to avoid confusion with other packages in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets' package and identifies it as a data frame. The original content of the dataset has not been modified.

Usage

data(USJudgeRatings_df)

Format

A data frame with 43 observations and 12 variables representing ratings for U.S. judges:

CONT

Numeric vector representing the judges' ratings on control.

INTG

Numeric vector representing the judges' ratings on integrity.

DMNR

Numeric vector representing the judges' ratings on demeanor.

DILG

Numeric vector representing the judges' ratings on diligence.

CFMG

Numeric vector representing the judges' ratings on communications with clients.

DECI

Numeric vector representing the judges' ratings on decisiveness.

PREP

Numeric vector representing the judges' ratings on preparation.

FAMI

Numeric vector representing the judges' ratings on family law expertise.

ORAL

Numeric vector representing the judges' ratings on oral communications.

WRIT

Numeric vector representing the judges' ratings on written communications.

PHYS

Numeric vector representing the judges' ratings on physical appearance.

RTEN

Numeric vector representing the judges' ratings on overall rating.

Source

U.S. judge ratings data.