Housing prices in Ames, Iowa

The dataset name has been changed to 'ames_tbl_df' to avoid confusion with other packages in the R ecosystem from which datasets have been sourced. This naming convention helps distinguish this dataset as part of the 'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original content of the dataset has not been modified in any way.

Usage

data(ames_tbl_df)

Format

A tibble with 2,930 observations and 82 variables:

Order

Row number in the dataset.

PID

Parcel Identifier.

area

Total house area in square feet.

price

Sale price of the house.

MS.SubClass

Building class type.

MS.Zoning

Zoning classification of the property.

Lot.Frontage

Lot frontage length in feet.

Lot.Area

Total lot area in square feet.

Street

Street type access to the property.

Alley

Alley type access.

Lot.Shape

Shape of the lot.

Land.Contour

Land contour around the property.

Utilities

Availability of utilities.

Lot.Config

Lot configuration.

Land.Slope

Slope of the land.

Neighborhood

Neighborhood in Ames.

Condition.1

Proximity to main conditions like railroads.

Condition.2

Proximity to secondary conditions.

Bldg.Type

Type of building.

House.Style

Architectural style of the house.

Overall.Qual

Overall quality of the materials and finish.

Overall.Cond

Overall condition of the house.

Year.Built

Year the house was built.

Year.Remod.Add

Year of the last remodel or addition.

Roof.Style

Roof style.

Roof.Matl

Roof material.

Exterior.1st

Primary exterior material.

Exterior.2nd

Secondary exterior material.

Mas.Vnr.Type

Masonry veneer type.

Mas.Vnr.Area

Masonry veneer area in square feet.

Exter.Qual

Exterior material quality.

Exter.Cond

Condition of the exterior material.

Foundation

Type of foundation.

Bsmt.Qual

Basement quality.

Bsmt.Cond

Basement condition.

Bsmt.Exposure

Basement exposure to the outside.

BsmtFin.Type.1

Type 1 of finished basement.

BsmtFin.SF.1

Square feet of finished basement type 1.

BsmtFin.Type.2

Type 2 of finished basement.

BsmtFin.SF.2

Square feet of finished basement type 2.

Bsmt.Unf.SF

Unfinished basement area in square feet.

Total.Bsmt.SF

Total basement area in square feet.

Heating

Type of heating system.

Heating.QC

Heating system quality.

Central.Air

Presence of central air conditioning.

Electrical

Type of electrical system.

X1st.Flr.SF

First floor area in square feet.

X2nd.Flr.SF

Second floor area in square feet.

Low.Qual.Fin.SF

Low-quality finished area in square feet.

Bsmt.Full.Bath

Number of full bathrooms in the basement.

Bsmt.Half.Bath

Number of half bathrooms in the basement.

Full.Bath

Number of full bathrooms above ground.

Half.Bath

Number of half bathrooms above ground.

Bedroom.AbvGr

Number of bedrooms above ground.

Kitchen.AbvGr

Number of kitchens above ground.

Kitchen.Qual

Kitchen quality.

TotRms.AbvGrd

Total number of rooms above ground.

Functional

Functionality of the house.

Fireplaces

Number of fireplaces.

Fireplace.Qu

Fireplace quality.

Garage.Type

Type of garage.

Garage.Yr.Blt

Year the garage was built.

Garage.Finish

Garage finish type.

Garage.Cars

Number of cars the garage can accommodate.

Garage.Area

Garage area in square feet.

Garage.Qual

Garage quality.

Garage.Cond

Garage condition.

Paved.Drive

Indicates whether the driveway is paved.

Wood.Deck.SF

Wood deck area in square feet.

Open.Porch.SF

Open porch area in square feet.

Enclosed.Porch

Enclosed porch area in square feet.

X3Ssn.Porch

Three-season porch area in square feet.

Screen.Porch

Screened porch area in square feet.

Pool.Area

Pool area in square feet.

Pool.QC

Pool quality.

Fence

Type of fence.

Misc.Feature

Miscellaneous features of the property.

Misc.Val

Value of miscellaneous features.

Mo.Sold

Month the house was sold.

Yr.Sold

Year the house was sold.

Sale.Type

Type of sale.

Sale.Condition

Condition of the sale.

Source

Ames Housing Dataset, provided by Dean De Cock