Batter Statistics for 2018 Major League Baseball (MLB) Season

The dataset name has been changed to 'mlb_players_18_tbl_df' to avoid confusion with other packages in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets' package and identifies it as a tibble. The original content of the dataset has not been modified.

Usage

data(mlb_players_18_tbl_df)

Format

A tibble with 1270 observations and 19 variables:

name

Character string representing the name of the player.

team

Character string indicating the team the player belongs to.

position

Character string indicating the position played by the player.

games

Integer representing the number of games played.

AB

Integer indicating the number of at-bats.

R

Integer representing the number of runs scored.

H

Integer representing the number of hits.

doubles

Integer indicating the number of doubles hit.

triples

Integer indicating the number of triples hit.

HR

Integer representing the number of home runs hit.

RBI

Integer indicating the number of runs batted in.

walks

Integer indicating the number of walks received.

strike_outs

Integer indicating the number of strikeouts.

stolen_bases

Integer representing the number of stolen bases.

caught_stealing_base

Integer indicating the number of times caught stealing.

AVG

Numeric representing the batting average.

OBP

Numeric representing the on-base percentage.

SLG

Numeric representing the slugging percentage.

OPS

Numeric representing the on-base plus slugging percentage.

Source

Data from Major League Baseball (MLB) player statistics for the 2018 season